% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two vectors}
\usage{
add(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric vector
}
\description{
This function adds two numeric vectors and returns
a numeric vector
}
\examples{
add(2,3)
add(mtcars$mpg, mtcars$hp)
}
