\name{deconv0}
\alias{deconv0}
\title{Deconvolution using fft}
\description{
Deconvolute two given sequences.
}
\usage{deconv0(z, x)}
\arguments{
  \item{z}{Response sequence to be deconvoluted}
  \item{x}{Input or Unit response sequence}
}
\details{
The lengthes of two sequences should be same. The order matters.
}
\value{
  Deconvolution result.
}
\author{Kyun-Seop Bae <k@acr.kr>}


