\name{Df}
\alias{Df}
\title{Density of F distribution}
\description{
Calculate the density of a F distribution
}
\usage{Df(f, nu1, nu2)}
\arguments{
  \item{f}{f value for an F distribution}
  \item{nu1}{the shape parameter nu1 of an F distribution}
  \item{nu2}{the shape parameter nu2 of an F distribution}
}
\details{
It calculates the density for f value of an F distribution.
}
\value{
  density for f
}
\author{Kyun-Seop Bae <k@acr.kr>}
