% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{An example dataset for demonstrating the functions in mathR}
\format{
\subsection{\code{example_data}}{

A data frame with 9 rows and 5 columns:
\describe{
\item{subject_id}{An integer denoting the subject id}
\item{modular}{The value for m to mod the function of interest by}
\item{maxdeg2_func}{Functions with a maximum degree of 2}
\item{param_graph}{Parameters for the line_line_intersection function}
\item{rand_func}{Random functions with any degree}
}
}
}
\usage{
example_data
}
\description{
An example dataset for demonstrating the functions in mathR
}
\keyword{datasets}
