% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_line_intersection.R
\name{line_line_intersection}
\alias{line_line_intersection}
\title{Plot intersections between two polynomial functions}
\usage{
line_line_intersection(a1, a2, a3, a4, b1, b2, b3, b4, interval)
}
\arguments{
\item{a1}{A numeric value for the coefficient of \eqn{x^3} for function 1}

\item{a2}{A numeric value for the coefficient of \eqn{x^2} for function 1}

\item{a3}{A numeric value for the coefficient of \eqn{x} for function 1}

\item{a4}{A numeric value for the constant coefficient for function 1}

\item{b1}{A numeric value for the coefficient of \eqn{x^3} for function 2}

\item{b2}{A numeric value for the coefficient of \eqn{x^2} for function 2}

\item{b3}{A numeric value for the coefficient of \eqn{x} for function 2}

\item{b4}{A numeric value for the constant coefficient for function 2}

\item{interval}{A numeric vector for the interested interval in \eqn{x}}
}
\value{
A \code{ggplot2} object displaying the two functions within the chosen
\eqn{x}-interval and the coordinates of intersecting points within that range.
}
\description{
Compute and plot the intersections between 2 polynomials (coefficient in the real number)
up to degree 3 over a specified interval.
}
\details{
The first polynomial is defined as

\deqn{
f_1(x) = a_1x^3 + a_2x^2 + a_3x + a_4
}

and the second polynomial is defined as

\deqn{
f_2(x) = b_1x^3 + b_2x^2 + b_3x + b_4.
}

To find the intersections, the function uses the package rootSolve
to numerically determine the roots of \eqn{f_1(x)-f_2(x)}. In other words, this operation
finds data points at which \eqn{f_1(x)=f_2(x)}.
}
\examples{

line_line_intersection(0,1,0,0,0,0,1,0,3)
line_line_intersection(0,0,5,0,0,0,5,0,3)
line_line_intersection(1,1,1,0,0,0,1,0,2)
}
