% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_arithmetic_for_function.R
\name{modular_arithmetic_for_function}
\alias{modular_arithmetic_for_function}
\title{Apply modular arithmetic to polynomials}
\usage{
modular_arithmetic_for_function(f1, m)
}
\arguments{
\item{f1}{A numeric vector of the coefficients of terms in the polynomial (in an order from the leading to the constant coefficient)}

\item{m}{A nonzero numeric value that the polynomial will be mod out by (modulo base)}
}
\value{
A character string representing the polynomial with coefficients reduced modulo \eqn{m}.
}
\description{
Apply modular arithmetic to the coefficients of a real polynomial and return the
resulting polynomial expression modulo \eqn{m}. This concept is heavily used in abstract algebra to reduce the
algebraic structures like rings and fields, and also in cryptography to restrict
polynomials to a finite field, allowing computers to process math without the coefficients growing
infinitely large.
}
\details{
Given a polynomial

\deqn{
f(x) = a_nx^n + a_{n-1}x^{n-1} + \dots + a_1x + a_0,
}

the function replaces each coefficient \eqn{a_i} with
\eqn{a_i \bmod m}.
}
\examples{

modular_arithmetic_for_function(c(1,3,4),4)
modular_arithmetic_for_function(c(1,3,1,7,5),3)
}
