% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply.R
\name{multiply}
\alias{multiply}
\title{Multiply two vectors}
\usage{
multiply(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric vector
}
\description{
This function multiplies two numeric vectors and returns
a numeric vector
}
\examples{
multiply(2,3)
multiply(mtcars$mpg, mtcars$hp)
}
