\name{NashSVD}
\alias{NashSVD}
\title{Singular Value Decomposition by Nash}
\description{
Singular value decomposition by Nash
}
\usage{NashSVD(M)}
\arguments{
  \item{M}{a symmetric matrix of real values}
}
\details{
This can be used for a least square problem.
}
\value{
  Decomposed result
}
\author{Kyun-Seop Bae <k@acr.kr>}
