% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quicksort.R
\name{quicksort}
\alias{quicksort}
\title{Quick Sort}
\usage{
quicksort(arr)
}
\arguments{
\item{arr}{a numeric vector to be sorted}
}
\description{
Sort a numeric vector in non-decreasing order
}
\examples{
#quicksort(c(9, 7, 1, 3, 6, -5))
#quicksort(c(9.8, 3, 4.0, 4.0, 9.8, -2.3))

}

