% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root_finding.R
\name{root_finding}
\alias{root_finding}
\title{Finding the roots of degree 1 and 2 polynomials with real number coefficients and 1 variable}
\usage{
root_finding(a, b, c)
}
\arguments{
\item{a}{A numeric value for the coefficient of x^2}

\item{b}{A numeric value for the coefficient of x}

\item{c}{A numeric value for the constant coefficient}
}
\value{
A numeric value of the roots or a message if the polynomial is invalid or irreducible
}
\description{
Finding the roots of degree 1 and 2 polynomials with real number coefficients and 1 variable
}
\examples{

root_finding(1,0,-9)
root_finding(4,4,1)
root_finding(0,0,5)
root_finding(25,100,1)
}
