\name{SQRT}
\alias{SQRT}
\title{Square-rooting}
\description{
Square-rooting of a real positive number. 
}
\usage{SQRT(x)}
\arguments{
  \item{x}{a real positive number to be square-rooted}
}
\details{
Ref. Plauger PJ The Standard C Library. 1992.
}
\value{
  square-rooted value
}
\author{Kyun-Seop Bae <k@acr.kr>}

