% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_conversion.R
\name{unit_conversion}
\alias{unit_conversion}
\title{Conversion from the U.S. customary length and weight units to SI units (meter and kilogram)}
\usage{
unit_conversion(x, unit)
}
\arguments{
\item{x}{A single positive numerical value}

\item{unit}{A U.S. customary unit type for length and weight}
}
\value{
A unit conversion of the input x in the U.S. customary unit to SI unit, meter for length and kilogram for weight
}
\description{
Conversion from the U.S. customary length and weight units to SI units (meter and kilogram)
}
\examples{
unit_conversion(120, 'lb')
}
