\name{conv0}
\alias{conv0}
\title{Convolution using fft}
\description{
Convolute two given sequences.
}
\usage{conv0(x, y)}
\arguments{
  \item{x}{Input sequence}
  \item{y}{Response sequence to unit impulse}
}
\details{
One is input and the other is unit response sequence. The lengthes of two sequences should be same. The order is not important
}
\value{
  convolution result.
}
\author{Kyun-Seop Bae <k@acr.kr>}


