\name{Dchisq}
\alias{Dchisq}
\title{Density of chi-square distribution}
\description{
Calculate the density of a chi-square distribution
}
\usage{Dchisq(x2, nu)}
\arguments{
  \item{x2}{chi-square value for a chi-square distribution}
  \item{nu}{the shape parameter nu of a chi-square distribution}
}
\details{
It calculates the density for chi-sqaure value of a chi-square distribution.
}
\value{
  density for x2
}
\examples{
Dchisq(3, 5)
dchisq(3, 5)
}
\author{Kyun-Seop Bae <k@acr.kr>}
