% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide.R
\name{divide}
\alias{divide}
\title{Dive one vector by another vector}
\usage{
divide(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric vector
}
\description{
This function divides one numeric vector by another
numeric vector and returns a numeric vector
}
\examples{
divide(2, 3)
divide(mtcars$mpg, mtcars$hp)
}
