\name{Dlnorm}
\alias{Dlnorm}
\title{Density of log-normal distribution}
\description{
Calculate the density of a log-normal distribution
}
\usage{Dlnorm(x, mu=0, sig=1)}
\arguments{
  \item{x}{x value for a log-normal distribution}
  \item{mu}{the shape parameter mu of a log-normal distribution}
  \item{sig}{the shape parameter sigma of a log-normal distribution}
}
\details{
It calculates the density for x value of a log-normal distribution.
}
\value{
  density for x
}
\author{Kyun-Seop Bae <k@acr.kr>}
