\name{Dnorm}
\alias{Dnorm}
\title{Density of normal distribution}
\description{
Calculate the density of a normal distribution
}
\usage{Dnorm(x, mu=0, sig=1)}
\arguments{
  \item{x}{x value for a normal distribution}
  \item{mu}{the shape parameter mu of a normal distribution}
  \item{sig}{the shape parameter sigma of a normal distribution}
}
\details{
It calculates the density for x value of a normal distribution.
}
\value{
  density for x
}
\author{Kyun-Seop Bae <k@acr.kr>}
