\name{EXP}
\alias{EXP}
\title{Exponential}
\description{
Exponential of a real value. 
}
\usage{EXP(x)}
\arguments{
  \item{x}{a real number to be Exponentiated}
}
\details{
Ref. Plauger PJ The Standard C Library. 1992.
}
\value{
  Exponential value of a real value.
}
\author{Kyun-Seop Bae <k@acr.kr>}

