% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_line_intersection.R
\name{line_line_intersection}
\alias{line_line_intersection}
\title{Finding the intersections between 2 polynomials (coefficient in the real nummber) up to degree 3 and plotting the intersections with the functions}
\usage{
line_line_intersection(a1, a2, a3, a4, b1, b2, b3, b4, interval)
}
\arguments{
\item{a1}{A numeric value for the coefficient of x^3 for function 1}

\item{a2}{A numeric value for the coefficient of x^2 for function 1}

\item{a3}{A numeric value for the coefficient of x for function 1}

\item{a4}{A numeric value for the constant coefficient for function 1}

\item{b1}{A numeric value for the coefficient of x^3 for function 2}

\item{b2}{A numeric value for the coefficient of x^2 for function 2}

\item{b3}{A numeric value for the coefficient of x for function 2}

\item{b4}{A numeric value for the constant coefficient for function 2}

\item{interval}{A numeric vector for the interested interval in x}
}
\value{
A plot of the functions within the chosen x-interval and the intersecting points within that range
}
\description{
Finding the intersections between 2 polynomials (coefficient in the real nummber) up to degree 3 and plotting the intersections with the functions
}
\examples{

line_line_intersection(0,1,0,0,0,0,1,0,3)
line_line_intersection(0,0,5,0,0,0,5,0,3)
line_line_intersection(1,1,1,0,0,0,1,0,2)
}
