\name{LOG}
\alias{LOG}
\title{Logarithm natural or common}
\description{
Natural or common log of a real positive value. 
}
\usage{LOG(x, DecFlag=FALSE)}
\arguments{
  \item{x}{a real positive number to be log-ed}
  \item{DecFlag}{FALSE for natural log, TRUE for common log}
}
\details{
Ref. Plauger PJ The Standard C Library. 1992.
}
\value{
  log value of base *e* or 10.
}
\author{Kyun-Seop Bae <k@acr.kr>}

