% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modular_arithmetic_for_function.R
\name{modular_arithmetic_for_function}
\alias{modular_arithmetic_for_function}
\title{Applying modular arithmetic to polynomials with coefficients in the real number}
\usage{
modular_arithmetic_for_function(f1, m)
}
\arguments{
\item{f1}{A numeric vector of the coefficients of terms in the polynomial (in an order from the leading to the constant coefficient)}

\item{m}{A numeric value that the polynomial will be mod out by (modulo base)}
}
\value{
A string of the polynomial modulo m
}
\description{
Applying modular arithmetic to polynomials with coefficients in the real number
}
\examples{

modular_arithmetic_for_function(c(1,3,4),4)
modular_arithmetic_for_function(c(1,3,1,7,5),3)
}
