\name{OuterProd}
\alias{OuterProd}
\title{Outer product of two three-dimensional vectors}
\description{
Calculate the outer product of two three-dimentional vectors
}
\usage{OuterProd(a, b)}
\arguments{
  \item{a}{the first vector}
  \item{b}{the second vector}
}
\details{
Outer product is not communtative.
}
\value{
  Outer product of two three-dimension vectors
}
\author{Kyun-Seop Bae <k@acr.kr>}
