\name{Ppois}
\alias{Ppois}
\title{Cumulative probability of Poisson distribution}
\description{
Calculate the cumulative probability P(X <= n) of a Poisson distribution
}
\usage{Ppois(n, lam)}
\arguments{
  \item{n}{count n for a Poisson distribution}
  \item{lam}{shape parameter lambda of a Poisson distribution}
}
\details{
It calculates the cumulative probability for n of a Poisson distribution.
}
\value{
  cumulative probability till n
}
\examples{
Ppois(0, 4)                  # P(X <= 0) = exp(-4)
Ppois(3, 4)
ppois(3, 4)
}
\author{Kyun-Seop Bae <k@acr.kr>}
