\name{Pt}
\alias{Pt}
\title{Cumulative probability of t distribution}
\description{
Calculate the cumulative probability for t of a t distribution
}
\usage{Pt(t, nu, mu=0, sig=1)}
\arguments{
  \item{t}{t value for a t distribution}
  \item{nu}{degree of freedom}
  \item{mu}{the shape parameter mu of a t distribution}
  \item{sig}{the shape parameter sigma of a t distribution}
}
\details{
It calculates the cumulative probability for t of a t distribution.
}
\value{
  cumulative probability till t
}
\author{Kyun-Seop Bae <k@acr.kr>}
