\name{Qf}
\alias{Qf}
\title{Reverse of Pf function}
\description{
Calculate the f for the cumulative probability p of an F distribution
}
\usage{Qf(p, nu1, nu2)}
\arguments{
  \item{p}{cumulative probability for f of an F distribution}
  \item{nu1}{the shape parameter nu1 of an F distribution}
  \item{nu2}{the shape parameter nu2 of an F distribution}
}
\details{
It calculates the f for the cumulative probability p of an F distribution.
}
\value{
  f of \code{Pf(f, nu1, nu2)}
}
\author{Kyun-Seop Bae <k@acr.kr>}
