\name{Qpois}
\alias{Qpois}
\title{Reverse of Ppois function}
\description{
Calculate the n for the cumulative probability p of a Poisson distribution
}
\usage{Qpois(p, lam)}
\arguments{
  \item{p}{cumulative proability of a Poisson distribution}
  \item{lam}{shape parameter lambda of a Poisson distribution}
}
\details{
It calculates the n for the cumulative probability p of a Poisson distribution.
}
\value{
  n of \code{Ppois(n, lam)}
}
\author{Kyun-Seop Bae <k@acr.kr>}
