\name{Rbeta}
\alias{Rbeta}
\title{Random deviate from beta distribution}
\description{
Generate random numbers from a beta distribution
}
\usage{Rbeta(n, alph, bet)}
\arguments{
  \item{n}{count, how many do you want}
  \item{alph}{shape parameter alpha of a beta distribution}
  \item{bet}{shape parameter beta of a beta distribution}
}
\details{
It generates random numbers from random numbers of other ditribution.
}
\value{
  a vector of random numbers
}
\author{Kyun-Seop Bae <k@acr.kr>}
