% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract.R
\name{subtract}
\alias{subtract}
\title{Subtract two vectors}
\usage{
subtract(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a numeric vector
}
\description{
This function subtracts two numeric vectors and returns
a numeric vector
}
\examples{
subtract(2, 3)
subtract(mtcars$mpg, mtcars$hp)
}
