\name{SymDet}
\alias{SymDet}
\title{Determinant of a Symmetric Matrix}
\description{
Determinant of a symmetric matrix
}
\usage{SymDet(SymMat)}
\arguments{
  \item{SymMat}{a symmetric matrix of real values}
}
\details{
Input should be a symmetric matrix of real values.
}
\value{
  Determinant of the input
}
\author{Kyun-Seop Bae <k@acr.kr>}
