\name{SymInv}
\alias{SymInv}
\title{Inverse of a Symmetric Matrix}
\description{
Inverse of a symmetric matrix
}
\usage{SymInv(SymMat)}
\arguments{
  \item{SymMat}{a symmetric matrix of real values}
}
\details{
Input should be a symmetric matrix of real values.
}
\value{
  Inverse of the input
}
\author{Kyun-Seop Bae <k@acr.kr>}
