\name{tableFactorial}
\alias{tableFactorial}
\title{Factorial using Table}
\description{
This function returns factorial value from the stored table. 
}
\usage{tableFactorial(n)}
\arguments{
  \item{n}{positive integer to be calculated}
}
\details{
This function returns factorial value from the table known to be precise.
}
\value{
  factorial value
}
\author{Kyun-Seop Bae <k@acr.kr>}
\examples{
tableFactorial(0)
tableFactorial(170)
tableFactorial(171)
}


